Added Framework for Phase Sum calculations
Added functions 'getphasesum()', 'getk3()' and 'getindep()' to ABC notebook (Thanks Chris!).

The getphasesum()  function will calculate the phase sums within a given range of k values (kmin to kmax).

Currently, the function returns no errors but the 'test' figure is incorrectly plotting, so work will still need to be done on this section of the notebook in the future.